requirements.txt
uvicorn==0.27.1
FastAPI==0.110.0
PyTesseract==0.3.10
pdf2image==1.17.0
pillow==10.2.0
python-mutipart==0.0.9
PyYAML==6.0.1
kerasOCR==0.9.3
tensorflow==2.15.0
keras==2.15.0
jinja2==3.1.3